split useWorldUsersById / useWorldUsersByIdWorkaround into it's own file + add @deprecated note
