+++ src/hooks/users/index.ts
@@ -5,47 +5,18 @@ import { User } from "types/User";
 import { WithId } from "utils/id";
 import { normalizeTimestampToMilliseconds } from "utils/time";
 
-import { worldUsersByIdSelector } from "utils/selectors";
-
 import { useConnectCurrentVenueNG } from "hooks/useConnectCurrentVenueNG";
-import { isLoaded } from "hooks/useFirestoreConnect";
 import { useUserLastSeenThreshold } from "hooks/useUserLastSeenThreshold";
-import { useSelector } from "hooks/useSelector";
 import { useVenueId } from "hooks/useVenueId";
 
-import { useConnectWorldUsers } from "./useConnectWorldUsers";
 import { useWorldUsers } from "./useWorldUsers";
 
 export { useConnectWorldUsers } from "./useConnectWorldUsers";
 export { useWorldUsers } from "./useWorldUsers";
-
-// @debt typing, Record implies that a User will exist for literally any given string, which is untrue
-export const useWorldUsersById = () => {
-  useConnectWorldUsers();
-
-  const worldUsersById = useSelector(worldUsersByIdSelector);
-
-  return useMemo(
-    () => ({
-      worldUsersById: worldUsersById ?? {},
-      isWorldUsersLoaded: isLoaded(worldUsersById),
-    }),
-    [worldUsersById]
-  );
-};
-
-// @debt typing, this uses Partial<Record<K,T>> to work around the bug where Record implies that a User will exist for literally any given string, which is untrue
-// @debt refactor all usages of Record<string, User> to Partial<Record<string, User>> (then we can remove this workaround)
-export const useWorldUsersByIdWorkaround = () => {
-  const {
-    worldUsersById: _worldUsersById,
-    isWorldUsersLoaded,
-  } = useWorldUsersById();
-
-  const worldUsersById: Partial<Record<string, User>> = _worldUsersById;
-
-  return { worldUsersById, isWorldUsersLoaded };
-};
+export {
+  useWorldUsersById,
+  useWorldUsersByIdWorkaround,
+} from "./useWorldUsersById";
 
 export const useRecentWorldUsers = (): {
   recentWorldUsers: readonly WithId<User>[];
